# docs/swagger.yaml
definitions:
...
  models.AuthInputS:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  models.UserInputS:
    properties:
      email:
        type: string
      location:
        $ref: '#/definitions/models.UserLocation'
      name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    type: object
  models.UserLocation:
    properties:
      coordinates:
        description: 'Example: [123.45, 67.89]'
        items:
          type: number
        type: array
      type:
        description: Coordinates is an array of two float numbers.
        type: string
    type: object
  models.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      location:
        $ref: '#/definitions/models.UserLocation'
      name:
        type: string
      phone_number:
        type: string
    type: object
  models.UserUpdateSwagger:
    properties:
      location:
        $ref: '#/definitions/models.UserLocation'
      name:
        type: string
      phone_number:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /users:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all users with their locations in GeoJSON format
      produces:
      - application/json
      responses:
        "200":
          description: A list of users
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: User successfully deleted
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Delete a user
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Updates details of an existing user by ID.
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserUpdateSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user details
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: Validation Error
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
      summary: Update user details
      tags:
      - users
  /users/authentication:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a token
      parameters:
      - description: User credentials for authentication
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.AuthInputS'
      produces:
      - application/json
      responses:
        "200":
          description: id, token" "ID and token of the authenticated user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Incorrect password or validation error
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Authenticate a user
      tags:
      - users
  /users/registration:
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided information
      parameters:
      - description: User data for registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserInputS'
      produces:
      - application/json
      responses:
        "200":
          description: id, token" "ID and token of the newly registered user
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Validation Error
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new user
      tags:
      - users
swagger: "2.0"